+++ classes/event/finding_created.php
@@ -1,0 +1,68 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+namespace local_handbook\event;
+
+/**
+ * Event: a quality finding was created (by a reader, an AI tool or an audit).
+ *
+ * @package   local_handbook
+ * @copyright Educación Helvética SA / EuropaSchule
+ * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class finding_created extends \core\event\base {
+
+    /**
+     * Initialise event data.
+     *
+     * @return void
+     */
+    protected function init() {
+        $this->data['crud'] = 'c';
+        $this->data['edulevel'] = self::LEVEL_OTHER;
+        $this->data['objecttable'] = 'local_handbook_finding';
+    }
+
+    /**
+     * Localised event name.
+     *
+     * @return string
+     */
+    public static function get_name() {
+        return get_string('reportproblem', 'local_handbook');
+    }
+
+    /**
+     * Event description for logs.
+     *
+     * @return string
+     */
+    public function get_description() {
+        $type = $this->other['findingtype'] ?? '';
+        $source = $this->other['source'] ?? '';
+        return "The user with id '{$this->userid}' created quality finding '{$this->objectid}' "
+            . "(type '{$type}', source '{$source}').";
+    }
+
+    /**
+     * Relevant URL.
+     *
+     * @return \moodle_url
+     */
+    public function get_url() {
+        return new \moodle_url('/local/handbook/manage/findings.php');
+    }
+}